# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage Equipment",
        "description": "Description Manage Equipment",
        "version": "0.1"
    },
    "paths": {
        "/api/equipments": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Get all equipments",
                "operationId": "getAllEquipments",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Equipment"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipments not found"
                    }
                }
            }
        },
        "/api/equipment/{id}": {
            "get": {
                "summary": "Retrieve information about an equipment",
                "description": "This method is called to retrieve information about an equipment",
                "operationId": "getDetailEquipment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipment/delete/{id}": {
            "delete": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Delete an equipment",
                "operationId": "30ac9840f2c556d5d232dcaac2226a9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Equipment deleted successfully"
                    },
                    "404": {
                        "description": "Equipment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Equipment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipment/create": {
            "post": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Create a new equipment",
                "operationId": "60e5b6097b7ecb0732c95ed861c45e5e",
                "requestBody": {
                    "description": "Equipment data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Equipment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipment created successfully",
                        "headers": {
                            "Location": {
                                "description": "URL of the created equipment",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipment/update/{id}": {
            "put": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Update an existing equipment",
                "operationId": "6d4d869fc4f16113bea50f0efe5dfb7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated equipment data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Equipment"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Equipment updated successfully"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Equipment": {
                "required": [
                    "name",
                    "category",
                    "number",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Equipment Name"
                    },
                    "category": {
                        "type": "string",
                        "example": "Equipment Category"
                    },
                    "number": {
                        "type": "string",
                        "example": "E12345"
                    },
                    "description": {
                        "type": "string",
                        "example": "Equipment description"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01T12:00:00+00:00"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-02T12:30:00+00:00"
                    }
                },
                "type": "object"
            }
        }
    }
}